extraer vista de venta de urls
